Keep module_path! calls under lib
